agent/requirements.txt
livekit-agents>=0.10.0
livekit-plugins-openai>=0.10.0
aiohttp>=3.10.0
python-dotenv>=1.0.1
